requirements.txt
asgiref==3.4.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
coverage==6.2
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
factory-boy==3.2.1
Faker==12.0.0
flake8==4.0.1
gunicorn==20.1.0
importlib-metadata==4.10.0
kafka-python==2.0.2
Markdown==3.3.6
mccabe==0.6.1
numpy==1.22.1
opencv-python-headless==4.5.5.62
Pillow==9.0.0
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tqdm==4.62.3
zipp==3.7.0
